requirements.txt
streamlit
panda
numpy
scikit-learn
plotly
joblib
matplotlib
seaborn
folium
streamlit-folium
pyttsx3
SpeechRecognition
pyaudio
openai
transformers
torch
langchain
streamlit-chat
geopy
requests
openmeteo-py
streamlit-option-menu
streamlit-aggrid
deep-translator